backend/requirement.txt
Flask==2.0.2
numpy==1.21.2
tensorflow==2.6.0
opencv-python==4.5.3.56
Pillow==8.3.2
keras==2.6.0
